To add codecoverage testing scenario
