+++ config/spock/SpockConfig.groovy
@@ -1,0 +1,10 @@
+import org.spockframework.runtime.model.parallel.ExecutionMode
+
+runner {
+  parallel {
+    enabled true
+    defaultSpecificationExecutionMode ExecutionMode.CONCURRENT
+    defaultExecutionMode ExecutionMode.SAME_THREAD
+    // fixed(3)
+  }
+}
