+++ src/test/groovy/com/github/prokod/gradle/crossbuild/CrossBuildPluginExtensionTest.groovy
@@ -174,9 +174,9 @@ dependencies {
         Assume.assumeTrue(testMavenCentralAccess())
         def result = GradleRunner.create()
                 .withGradleVersion(gradleVersion)
-                .withProjectDir(dir.root)
+                .withProjectDir(dir.toFile())
                 .withPluginClasspath()
-                .withDebug(true)
+                /*@withDebug@*/
                 .withArguments('crossBuildResolvedDsl', '--info', '--stacktrace')
                 .build()
