+++ src/test/groovy/com/github/prokod/gradle/crossbuild/CrossBuildPluginPublishingTest.groovy
@@ -21,6 +21,8 @@ import org.skyscreamer.jsonassert.JSONAssert
 import org.xmlunit.diff.Diff
 import spock.lang.Unroll
 
+import java.nio.file.Files
+
 import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
 class CrossBuildPluginPublishingTest extends CrossBuildGradleRunnerSpec {
@@ -110,14 +112,14 @@ subprojects {
         publishing {
             publications {
                 crossBuildSpark230_211(MavenPublication) {
-                    ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                    afterEvaluate {
                         artifact crossBuildSpark230_211Jar
-                    ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+                    }
                 }
                 crossBuildSpark240_212(MavenPublication) {
-                    ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                    afterEvaluate {
                         artifact crossBuildSpark240_212Jar
-                    ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+                    }
                 }
             }
         }
@@ -238,9 +240,9 @@ crossBuild {
 publishing {
     publications {
         crossBuildSpark230_211(MavenPublication) {
-            ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+            afterEvaluate {
                 artifact crossBuildSpark230_211Jar
-            ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+            }
         }
     }
 }
@@ -254,9 +256,9 @@ dependencies {
         Assume.assumeTrue(testMavenCentralAccess())
         def result = GradleRunner.create()
                 .withGradleVersion(gradleVersion)
-                .withProjectDir(dir.root)
+                .withProjectDir(dir.toFile())
                 .withPluginClasspath()
-                .withDebug(true)
+                /*@withDebug@*/
                 .withArguments('crossBuildResolvedConfigs', 'publishToMavenLocal', '--info', '--stacktrace')
                 .build()
 
@@ -264,12 +266,12 @@ dependencies {
         result.task(":app:crossBuildSpark230_211Jar").outcome == SUCCESS
         result.task(":app:crossBuildResolvedConfigs").outcome == SUCCESS
 
-        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.11*.jar")
-        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.12*.jar")
-        fileExists("$dir.root.absolutePath/lib2/build/libs/lib2_2.11*.jar")
-        fileExists("$dir.root.absolutePath/lib2/build/libs/lib2_2.12*.jar")
-        fileExists("$dir.root.absolutePath/app/build/libs/app_2.11*.jar")
-        !fileExists("$dir.root.absolutePath/app/build/libs/app_2.12*.jar")
+        fileExists(dir.resolve('lib/build/libs/lib_2.11*.jar'))
+        fileExists(dir.resolve('lib/build/libs/lib_2.12*.jar'))
+        fileExists(dir.resolve('lib2/build/libs/lib2_2.11*.jar'))
+        fileExists(dir.resolve('lib2/build/libs/lib2_2.12*.jar'))
+        fileExists(dir.resolve('app/build/libs/app_2.11*.jar'))
+        !fileExists(dir.resolve('app/build/libs/app_2.12*.jar'))
 
         when:
         def expectedJsonAsText = loadResourceAsText(dsv: defaultScalaVersion,
@@ -285,15 +287,15 @@ dependencies {
         JSONAssert.assertEquals(expectedJsonAsText, actualJsonAsText, false)
 
         when:
-        def pom211File = new File("${dir.root.absolutePath}${File.separator}lib${File.separator}build${File.separator}generated-pom_2.11.xml")
-        def pom212File = new File("${dir.root.absolutePath}${File.separator}lib${File.separator}build${File.separator}generated-pom_2.12.xml")
+        def pom211Path = dir.resolve("lib${File.separator}build${File.separator}generated-pom_2.11.xml")
+        def pom212Path = dir.resolve("lib${File.separator}build${File.separator}generated-pom_2.12.xml")
 
         then:
-        pom211File.exists()
-        pom212File.exists()
+        Files.exists(pom211Path)
+        Files.exists(pom212Path)
 
         when:
-        def pom211 = new XmlSlurper().parse(pom211File)
+        def pom211 = new XmlSlurper().parse(pom211Path.toFile())
 
         then:
         pom211.dependencies.dependency.size() == 2
@@ -307,7 +309,7 @@ dependencies {
         pom211.dependencies.dependency[1].scope == 'runtime'
 
         when:
-        def pom212 = new XmlSlurper().parse(pom212File)
+        def pom212 = new XmlSlurper().parse(pom212Path.toFile())
 
         then:
         pom212.dependencies.dependency.size() == 2
@@ -321,38 +323,38 @@ dependencies {
         pom211.dependencies.dependency[1].scope == 'runtime'
 
         when:
-        def lib2pom211File = new File("${dir.root.absolutePath}${File.separator}lib2${File.separator}build${File.separator}generated-pom_2.11.xml")
-        def lib2pom212File = new File("${dir.root.absolutePath}${File.separator}lib2${File.separator}build${File.separator}generated-pom_2.12.xml")
+        def lib2pom211Path = dir.resolve("lib2${File.separator}build${File.separator}generated-pom_2.11.xml")
+        def lib2pom212Path = dir.resolve("lib2${File.separator}build${File.separator}generated-pom_2.12.xml")
 
         then:
-        lib2pom211File.exists()
-        lib2pom212File.exists()
+        Files.exists(lib2pom211Path)
+        Files.exists(lib2pom212Path)
 
         when:
         def expectedLib2Pom211 = loadResourceAsText(sv: '2.11', csv: '2.11.12', '/04-pom_lib2-00.xml')
-        Diff d211 = pomDiffFor(expectedLib2Pom211, lib2pom211File)
+        Diff d211 = pomDiffFor(expectedLib2Pom211, lib2pom211Path.toFile())
 
         then:
         !d211.hasDifferences()
 
         when:
         def expectedLib2Pom212 = loadResourceAsText(sv: '2.12', csv: '2.12.8','/04-pom_lib2-00.xml')
-        Diff d212 = pomDiffFor(expectedLib2Pom212, lib2pom212File)
+        Diff d212 = pomDiffFor(expectedLib2Pom212, lib2pom212Path.toFile())
 
         then:
         !d212.hasDifferences()
 
         when:
-        def appPom211File = new File("${dir.root.absolutePath}${File.separator}app${File.separator}build${File.separator}generated-pom_2.11.xml")
-        def appPom212File = new File("${dir.root.absolutePath}${File.separator}app${File.separator}build${File.separator}generated-pom_2.12.xml")
+        def appPom211Path = dir.resolve("app${File.separator}build${File.separator}generated-pom_2.11.xml")
+        def appPom212Path = dir.resolve("app${File.separator}build${File.separator}generated-pom_2.12.xml")
 
         then:
-        appPom211File.exists()
-        !appPom212File.exists()
+        Files.exists(appPom211Path)
+        !Files.exists(appPom212Path)
 
         when:
         def expectedAppPom211 = loadResourceAsText( '/04-pom_app-00.xml')
-        Diff dApp211 = pomDiffFor(expectedAppPom211, appPom211File)
+        Diff dApp211 = pomDiffFor(expectedAppPom211, appPom211Path.toFile())
 
         then:
         !dApp211.hasDifferences()
